+++ src/Listeners/StopValidationListener.php
@@ -1,0 +1,89 @@
+<?php
+
+/**
+ * TechDivision\Import\Listeners\StopValidationListener
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Listeners;
+
+use League\Event\EventInterface;
+use League\Event\AbstractListener;
+use TechDivision\Import\Utils\RegistryKeys;
+use TechDivision\Import\ApplicationInterface;
+use TechDivision\Import\Services\RegistryProcessorInterface;
+use TechDivision\Import\Subjects\SubjectInterface;
+use TechDivision\Import\Plugins\PluginInterface;
+
+/**
+ * A listener implementation that stops the application on validation errors.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2019 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import
+ * @link      http://www.techdivision.com
+ */
+class StopValidationListener extends AbstractListener
+{
+
+    /**
+     * The import processor instance.
+     *
+     * @var \TechDivision\Import\Services\RegistryProcessorInterface
+     */
+    protected $registryProcessor;
+
+    /**
+     * Initializes the plugin with the application instance.
+     *
+     * @param \TechDivision\Import\Services\RegistryProcessorInterface $registryProcessor The registry processor instance
+     */
+    public function __construct( RegistryProcessorInterface $registryProcessor)
+    {
+        $this->registryProcessor = $registryProcessor;
+    }
+
+    /**
+     * Handle the event.
+     *
+     * @param \League\Event\EventInterface                      $event  The event that triggered the listener
+     * @param \TechDivision\Import\Plugins\PluginInterface|null $plugin The plugin instance
+     *
+     * @return void
+     */
+    public function handle(EventInterface $event, PluginInterface $plugin = null)
+    {
+
+        // load the status from the registry
+        $status = $this->getRegistryProcessor()->getAttribute(RegistryKeys::STATUS);
+
+        // query whether or not we've validation errors
+        if (isset($status[RegistryKeys::VALIDATIONS]) && sizeof($status[RegistryKeys::VALIDATIONS]) > 0) {
+            $plugin->getApplication()->stop('Stopped processing because of validation errors');
+        }
+    }
+
+    /**
+     * Return's the registry processor instance.
+     *
+     * @return \TechDivision\Import\Services\RegistryProcessorInterface The registry processor instance
+     */
+    protected function getRegistryProcessor()
+    {
+        return $this->registryProcessor;
+    }
+}